[hotfix][runtime] Extract common inner classes and methods to ExecutionGraphInfoStoreTestUtils
