+++ flink-runtime/src/test/java/org/apache/flink/runtime/dispatcher/FileExecutionGraphInfoStoreTest.java
@@ -21,42 +21,24 @@ package org.apache.flink.runtime.dispatcher;
 import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.common.JobStatus;
 import org.apache.flink.api.common.time.Time;
-import org.apache.flink.configuration.Configuration;
-import org.apache.flink.core.testutils.OneShotLatch;
-import org.apache.flink.runtime.blob.BlobServer;
-import org.apache.flink.runtime.clusterframework.types.ResourceID;
-import org.apache.flink.runtime.entrypoint.component.DefaultDispatcherResourceManagerComponentFactory;
-import org.apache.flink.runtime.entrypoint.component.DispatcherResourceManagerComponent;
-import org.apache.flink.runtime.entrypoint.component.DispatcherResourceManagerComponentFactory;
-import org.apache.flink.runtime.execution.Environment;
 import org.apache.flink.runtime.executiongraph.ArchivedExecutionGraph;
-import org.apache.flink.runtime.heartbeat.HeartbeatServices;
-import org.apache.flink.runtime.highavailability.HighAvailabilityServices;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.jobgraph.JobGraphTestUtils;
 import org.apache.flink.runtime.jobgraph.JobVertex;
-import org.apache.flink.runtime.jobgraph.tasks.AbstractInvokable;
 import org.apache.flink.runtime.messages.webmonitor.JobDetails;
 import org.apache.flink.runtime.messages.webmonitor.JobsOverview;
-import org.apache.flink.runtime.metrics.MetricRegistry;
 import org.apache.flink.runtime.minicluster.MiniCluster;
 import org.apache.flink.runtime.minicluster.MiniClusterConfiguration;
-import org.apache.flink.runtime.resourcemanager.StandaloneResourceManagerFactory;
 import org.apache.flink.runtime.rest.handler.legacy.utils.ArchivedExecutionGraphBuilder;
-import org.apache.flink.runtime.rpc.FatalErrorHandler;
 import org.apache.flink.runtime.scheduler.ExecutionGraphInfo;
 import org.apache.flink.runtime.util.ManualTicker;
-import org.apache.flink.runtime.webmonitor.retriever.MetricQueryServiceRetriever;
 import org.apache.flink.testutils.TestingUtils;
-import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.TestLogger;
 import org.apache.flink.util.concurrent.ManuallyTriggeredScheduledExecutor;
 
 import org.apache.flink.shaded.guava30.com.google.common.base.Ticker;
 import org.apache.flink.shaded.guava30.com.google.common.cache.LoadingCache;
 
-import org.hamcrest.BaseMatcher;
-import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;
 import org.junit.ClassRule;
@@ -66,27 +48,20 @@ import org.junit.rules.TemporaryFolder;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
-import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
+import static org.apache.flink.runtime.dispatcher.ExecutionGraphInfoStoreTestUtils.createDefaultExecutionGraphInfoStore;
+import static org.apache.flink.runtime.dispatcher.ExecutionGraphInfoStoreTestUtils.generateJobDetails;
+import static org.apache.flink.runtime.dispatcher.ExecutionGraphInfoStoreTestUtils.generateTerminalExecutionGraphInfos;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 /** Tests for the {@link FileExecutionGraphInfoStore}. */
 public class FileExecutionGraphInfoStoreTest extends TestLogger {
-
-    private static final List<JobStatus> GLOBALLY_TERMINAL_JOB_STATUS =
-            Arrays.stream(JobStatus.values())
-                    .filter(JobStatus::isGloballyTerminalState)
-                    .collect(Collectors.toList());
-
     @ClassRule public static TemporaryFolder temporaryFolder = new TemporaryFolder();
 
     /**
@@ -346,157 +321,19 @@ public class FileExecutionGraphInfoStoreTest extends TestLogger {
     /** Tests that a session cluster can terminate gracefully when jobs are still running. */
     @Test
     public void testPutSuspendedJobOnClusterShutdown() throws Exception {
+        File rootDir = temporaryFolder.newFolder();
         try (final MiniCluster miniCluster =
-                new PersistingMiniCluster(new MiniClusterConfiguration.Builder().build())) {
+                new ExecutionGraphInfoStoreTestUtils.PersistingMiniCluster(
+                        new MiniClusterConfiguration.Builder().build(), rootDir)) {
             miniCluster.start();
             final JobVertex vertex = new JobVertex("blockingVertex");
             // The adaptive scheduler expects that every vertex has a configured parallelism
             vertex.setParallelism(1);
-            vertex.setInvokableClass(SignallingBlockingNoOpInvokable.class);
+            vertex.setInvokableClass(
+                    ExecutionGraphInfoStoreTestUtils.SignallingBlockingNoOpInvokable.class);
             final JobGraph jobGraph = JobGraphTestUtils.streamingJobGraph(vertex);
             miniCluster.submitJob(jobGraph);
-            SignallingBlockingNoOpInvokable.LATCH.await();
-        }
-    }
-
-    /**
-     * Invokable which signals with {@link SignallingBlockingNoOpInvokable#LATCH} when it is invoked
-     * and blocks forever afterwards.
-     */
-    public static class SignallingBlockingNoOpInvokable extends AbstractInvokable {
-
-        /** Latch used to signal an initial invocation. */
-        public static final OneShotLatch LATCH = new OneShotLatch();
-
-        public SignallingBlockingNoOpInvokable(Environment environment) {
-            super(environment);
-        }
-
-        @Override
-        public void invoke() throws Exception {
-            LATCH.trigger();
-            Thread.sleep(Long.MAX_VALUE);
-        }
-    }
-
-    private class PersistingMiniCluster extends MiniCluster {
-
-        PersistingMiniCluster(MiniClusterConfiguration miniClusterConfiguration) {
-            super(miniClusterConfiguration);
-        }
-
-        @Override
-        protected Collection<? extends DispatcherResourceManagerComponent>
-                createDispatcherResourceManagerComponents(
-                        Configuration configuration,
-                        RpcServiceFactory rpcServiceFactory,
-                        HighAvailabilityServices haServices,
-                        BlobServer blobServer,
-                        HeartbeatServices heartbeatServices,
-                        MetricRegistry metricRegistry,
-                        MetricQueryServiceRetriever metricQueryServiceRetriever,
-                        FatalErrorHandler fatalErrorHandler)
-                        throws Exception {
-            final DispatcherResourceManagerComponentFactory
-                    dispatcherResourceManagerComponentFactory =
-                            DefaultDispatcherResourceManagerComponentFactory
-                                    .createSessionComponentFactory(
-                                            StandaloneResourceManagerFactory.getInstance());
-
-            final File rootDir = temporaryFolder.newFolder();
-            final ExecutionGraphInfoStore executionGraphInfoStore =
-                    createDefaultExecutionGraphInfoStore(rootDir);
-
-            return Collections.singleton(
-                    dispatcherResourceManagerComponentFactory.create(
-                            configuration,
-                            ResourceID.generate(),
-                            getIOExecutor(),
-                            rpcServiceFactory.createRpcService(),
-                            haServices,
-                            blobServer,
-                            heartbeatServices,
-                            metricRegistry,
-                            executionGraphInfoStore,
-                            metricQueryServiceRetriever,
-                            fatalErrorHandler));
-        }
-    }
-
-    private Collection<ExecutionGraphInfo> generateTerminalExecutionGraphInfos(int number) {
-        final Collection<ExecutionGraphInfo> executionGraphInfos = new ArrayList<>(number);
-
-        for (int i = 0; i < number; i++) {
-            final JobStatus state =
-                    GLOBALLY_TERMINAL_JOB_STATUS.get(
-                            ThreadLocalRandom.current()
-                                    .nextInt(GLOBALLY_TERMINAL_JOB_STATUS.size()));
-            executionGraphInfos.add(
-                    new ExecutionGraphInfo(
-                            new ArchivedExecutionGraphBuilder().setState(state).build()));
-        }
-
-        return executionGraphInfos;
-    }
-
-    private FileExecutionGraphInfoStore createDefaultExecutionGraphInfoStore(File storageDirectory)
-            throws IOException {
-        return new FileExecutionGraphInfoStore(
-                storageDirectory,
-                Time.hours(1L),
-                Integer.MAX_VALUE,
-                10000L,
-                TestingUtils.defaultScheduledExecutor(),
-                Ticker.systemTicker());
-    }
-
-    private static final class PartialExecutionGraphInfoMatcher
-            extends BaseMatcher<ExecutionGraphInfo> {
-
-        private final ExecutionGraphInfo expectedExecutionGraphInfo;
-
-        private PartialExecutionGraphInfoMatcher(ExecutionGraphInfo expectedExecutionGraphInfo) {
-            this.expectedExecutionGraphInfo =
-                    Preconditions.checkNotNull(expectedExecutionGraphInfo);
-        }
-
-        @Override
-        public boolean matches(Object o) {
-            if (expectedExecutionGraphInfo == o) {
-                return true;
-            }
-            if (o == null || expectedExecutionGraphInfo.getClass() != o.getClass()) {
-                return false;
-            }
-            ExecutionGraphInfo that = (ExecutionGraphInfo) o;
-
-            ArchivedExecutionGraph thisExecutionGraph =
-                    expectedExecutionGraphInfo.getArchivedExecutionGraph();
-            ArchivedExecutionGraph thatExecutionGraph = that.getArchivedExecutionGraph();
-            return thisExecutionGraph.isStoppable() == thatExecutionGraph.isStoppable()
-                    && Objects.equals(thisExecutionGraph.getJobID(), thatExecutionGraph.getJobID())
-                    && Objects.equals(
-                            thisExecutionGraph.getJobName(), thatExecutionGraph.getJobName())
-                    && thisExecutionGraph.getState() == thatExecutionGraph.getState()
-                    && Objects.equals(
-                            thisExecutionGraph.getJsonPlan(), thatExecutionGraph.getJsonPlan())
-                    && Objects.equals(
-                            thisExecutionGraph.getAccumulatorsSerialized(),
-                            thatExecutionGraph.getAccumulatorsSerialized())
-                    && Objects.equals(
-                            thisExecutionGraph.getCheckpointCoordinatorConfiguration(),
-                            thatExecutionGraph.getCheckpointCoordinatorConfiguration())
-                    && thisExecutionGraph.getAllVertices().size()
-                            == thatExecutionGraph.getAllVertices().size()
-                    && Objects.equals(
-                            expectedExecutionGraphInfo.getExceptionHistory(),
-                            that.getExceptionHistory());
-        }
-
-        @Override
-        public void describeTo(Description description) {
-            description.appendText(
-                    "Matches against " + ExecutionGraphInfo.class.getSimpleName() + '.');
+            ExecutionGraphInfoStoreTestUtils.SignallingBlockingNoOpInvokable.LATCH.await();
         }
     }
 
@@ -521,20 +358,14 @@ public class FileExecutionGraphInfoStoreTest extends TestLogger {
 
             assertThat(
                     executionGraphStore.get(dummyExecutionGraphInfo.getJobId()),
-                    new PartialExecutionGraphInfoMatcher(dummyExecutionGraphInfo));
+                    new ExecutionGraphInfoStoreTestUtils.PartialExecutionGraphInfoMatcher(
+                            dummyExecutionGraphInfo));
         }
     }
 
     private static Matcher<ExecutionGraphInfo> matchesPartiallyWith(
             ExecutionGraphInfo executionGraphInfo) {
-        return new PartialExecutionGraphInfoMatcher(executionGraphInfo);
-    }
-
-    private static Collection<JobDetails> generateJobDetails(
-            Collection<ExecutionGraphInfo> executionGraphInfos) {
-        return executionGraphInfos.stream()
-                .map(ExecutionGraphInfo::getArchivedExecutionGraph)
-                .map(JobDetails::createDetailsForJob)
-                .collect(Collectors.toList());
+        return new ExecutionGraphInfoStoreTestUtils.PartialExecutionGraphInfoMatcher(
+                executionGraphInfo);
     }
 }
